Working on scorer

+++ should_i_run/lib/model/api_handler.dart
@@ -1,13 +1,13 @@
 class ApiHandler {
   double tempResult;
   double precipResult;
-  double airQualityResult;
+  int airQualityResult;
   double humidityResult;
 
   void init() {
     tempResult = 5.0;
     precipResult = 2.0;
-    airQualityResult = 2.0;
+    airQualityResult = 2;
     humidityResult = 20.0;
   }
 }
